Refactor tests into new directory layout
