+++ sdk/python/tests/example_repos/example_feature_repo_with_missing_bq_source.py
