+++ sdk/python/tests/integration/scaffolding/test_init.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta
 from pathlib import Path
 from textwrap import dedent
 
-from tests.cli_utils import CliRunner
+from tests.utils.cli_utils import CliRunner
 
 
 def test_repo_init() -> None:
